+++ glamm/shaders/blit_maps.fs
@@ -18,7 +18,12 @@ main()
     if (merged_map_color.r == 0.5) {
       fragment_color = vec4(map_color.rgb, 1.0);
     } else {
-      fragment_color = vec4(0.31, 0.78, 0.47, 1.0); // emerald greend
+      // fragment_color = vec4(0.31, 0.78, 0.47, 1.0); // emerald greend
+      if (map_color.r == 0.0) {
+        fragment_color = map_color;
+      } else {
+        fragment_color = merged_map_color;
+      }
     }
   } else {
     fragment_color = vec4(merged_map_color.rgb, 1.0);