+++ CHANGELOG.md
@@ -12,5 +12,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Created basic GLUT window for visual debugging
 - Implemented simple vertex and fragment shader for drawing a map
 - `DrawMapShader` supports rendering a `OccupancyGridTextureMap`
+- `BlitMapsShader` supports merging `OccupancyGridTextureMap`s
 
 [Unreleased]: https://github.com/jdelacroix/glamm/compare/v0.1.0...HEAD